Fixed the cursor size in the quit menu.

+++ src/menu/messagebox.cpp
@@ -212,12 +212,13 @@ void DMessageBoxMenu::Drawer ()
 		{
 			if ((DMenu::MenuTime%8) < 6)
 			{
+				// [BB] Using Skulltag's increased precision of CleanXfac/CleanYfac is wrong here.
 				screen->DrawText(ConFont, OptionSettings.mFontColorSelection,
-					(150 - 160) * CleanXfac + screen->GetWidth() / 2,
-					(y + (fontheight + 1) * messageSelection - 100 + fontheight/2 - 5) * CleanYfac + screen->GetHeight() / 2,
+					(150 - 160) * static_cast<int>(CleanXfac) + screen->GetWidth() / 2,
+					(y + (fontheight + 1) * messageSelection - 100 + fontheight/2 - 5) * static_cast<int>(CleanYfac) + screen->GetHeight() / 2,
 					"\xd",
-					DTA_CellX, 8 * CleanXfac,
-					DTA_CellY, 8 * CleanYfac,
+					DTA_CellX, 8 * static_cast<int>(CleanXfac),
+					DTA_CellY, 8 * static_cast<int>(CleanYfac),
 					TAG_DONE);
 			}
 		}
